net: dsa: Add generic port map datastructure
For situations where a subset of ports must be tracked, but using
linked lists is not practical, allow the set to be kept in a bitmap
instead.

As an example, tracking the members of a given MDB group is not
feasible using linked lists, as a port can be a member of an arbitrary
number of groups.