net: bridge: switchdev: Send mrouter updates per VLAN and protocol
Even before this change, drivers would receive notifications when a
port's or bridge's multicast router status changed, via the
SWITCHDEV_ATTR_ID_{BRIDGE,PORT}_MROUTER attributes. Unfortunately, the
semantics around these notifications are cumbersome for drivers:

While the bridge has knowledge of both the VLAN ID and
protocol (IPv4/6) for which the notification applies, this information
is hidden from the receiver. This prevents drivers from applying any
fine grained hardware multicast filtering capabilities it may
have. Some hardware may have separate flood vectors for unregistered
IPv4 and IPv6 for example, others may be able to control it per VLAN,
etc.

The situation is further complicated by the fact that, for legacy
reasons, these attributes are signaled per-VLAN (even though the VID
is not available), but not per protocol. For this reason we can't
simply add the VID and protocol information to the existing attributes
without breaking changes.

Therefore, introduce a new notification that contains the full
context. Choose the object abstraction, since the original boolean
attribute now expands into a dynamic list of (vid, proto) tuples.

Only send notifications on inactive-to-active transitions and vice
versa, not when bridge internal states change that do not affect the
data plane. As an example, if a port is currently marked as a router
port because the network's querier sits somewhere behind it, and then
a user configures that port as a permanent router port, then that
should not be signaled to the driver. Because while those are two
distinct states from the bridge's perspective, they are logically
equivalent to a driver. As a consequence of this, drivers rely on a
simple reference counter to determine if a port is a multicast router
in any context by tracking the difference between the number of
SWITCHDEV_OBJ_ID_MROUTER additions vs. deletions.