net: bridge: switchdev: Maintain an MDB of multicast router ports
Hardware that does not maintain a separate list of router ports, like
the bridge does, every entry of its MDB typically needs to contain the
union of (1) the members of that group and (2) all ports marked as
multicast routers for the VLAN ID and protocol of the group in
question.

Rather than each switchdev driver needing to create their own caching
layer to keep track of this information, provide a separate (router)
MDB which keeps track of the relevant multicast groups that need
updating whenever memberships and mrouter states change.

To make use of this, a driver can keep a simple reference counter for
each port membership - which is already needed for port and host MDB
entries, to keep track of multiple IP groups sharing the same L2
address - and simply keep the corresponding hardware bit set as long
as the reference count remains positive.