net: bridge: switchdev: Maintain an MDB of multicast router ports
In contrast to the bridge's software implementation of multicast
router ports, lots of hardware does not have a dedicated facility for
tracking these states. Instead, a multicast router port must typically
be implemented in two parts:

1. Including the port in the most specific flooding mask available -
   to make sure that all unregistered groups for which the router is
   responsible reaches it. Sometimes flooding can be limited per
   protocol, sometimes per VLAN, and as a last resort some hardware
   has to enable flooding of _all_ unregistered multicast on the port.

2. Injecting the port into every registered group entry which matches
   the router's VLAN ID and protocol - to ensure that the router
   receives all registered groups.

While (1) is typically very driver specific, much of the heavy lifting
of implementing (2) is driver independent, requiring access to state
that the bridge already tracks, which would otherwise have to be
duplicated in a caching layer in each driver.

Therefore, provide a router MDB, which satisfies the following
invariant:

For every registered group, belonging to a given VLAN and protocol in
either the port or host MDB, the router MDB will contain an entry for
the same group, with each router port matching the VLAN and protocol
as its members.

The invariant is upheld by reacting to two classes of events:

- Whenever a router port is added or deleted, send a router MDB
  notification to each port, in every registered matching group, in
  the port and host MDBs.

- Whenever a new membership is registered on a port, send router MDB
  notifications to that port, for each applicable router port.

Drivers wanting to make use of this information are expected to
reference count router MDB memberships in the same way it likely
already does with regular port and host entries (typically needed to
keep track of multiple IP group memberships sharing the same L2
address). By keeping the union of the port-, host- and router MDB in
its hardware table, drivers can ensure that:

- All registered groups will also always include all relevant router
  ports.

- No stale group entries consisting of only router ports will ever be
  loaded. I.e., once the last port- or host member leaves the group,
  the reference count for all applicable routers in that group will
  also hit zero. The group will thus be reclassified as unregistered,
  and whatever facility is used to implement (1) will take over
  responsibility for making sure that any traffic to that group still
  reaches all routers.