net: dsa: Let drivers opt-in to inject mrouter ports into their MDBs
When a driver so chooses, inject multicast router ports into the
switch's MDB alongside the registered members. This will allow
registered IP groups to reach routers on devices that have dedicated
facility to accomplish this task.

As with port MDB objects, we map foreign ports to the host MDB to make
sure that those groups reach the CPU, where they can be forwarded to
the member by the bridge.