ptp: zl80732: Fix adjustments with negative offset
When calling adjtime with a negative offset that was smaller than 1
second then it was calculating wrong register values. It was calculating
wrong because the offset was negative and then used div_u64_rem which
gives a wrong result when used with negative numbers. Actually this
div_u6_rem is not needed because we know already that the delta is
smaller than 1 second. Therefore the fix consists in removing this and
just use the delta in the rest of the function.