ptp: zl80732: Fix when enabling 1PPS
On each output there are 2 pins, P and N pins. Each of this pin
correspond to one ptp pin. Each output can be controlled individually
but the pins are for the output are controlled by same output. So it
possible to generate 1 PPS only on 1 pin and the other one not to be
enabled.
The issue was that whenever was generating a 1PPS then it would just say
the value of the PIN regardless of the value of the other pin, so it
could disable the 1PPS on the other pin.
This try to fix this by checking in which state are the pins and set the
correct value. Also when stoping the 1PPS make sure to stop the 1PPS for
all the outputs and not only for the first one. Also when adjusting the
time adjust the output on all the pins that are connected and not only
on the first one.