lan966x: Fix when adding inteface under a lag
There was a crash when running the command:
ip link set dev eth0 master bond0

The reason is because when adding a inteface under the lag it would go
through all the ports and try to figure out which other ports are under
that lag inteface. And the issue is that lan966x can have ports that are
NULL pointer as they are not probed. So then iterating over these ports
it would just crash.
The fix consists in actually checking for NULL pointers before accesing
something from the ports.