sparx5: Make sure that frames are at least 60
Make sure that the size of the frame is at least 60. The thing is when
APPL is injecting a frame, will add the IFH which is already 36 bytes to
the front of the frame. So the frame then needs to be at least 24 bytes.
But that is not correct, the frame needs to be at least 60 bytes so
extend the frame to the size of 60.