sparx5: Take tx lock when injecting frames
When transmitting frames from the CPU take a lock. It was observed that
when having 2 threads which try to inject the frames, then it would
break the DCB chain so it would get a closed loop of DCBs so then the HW
will transmit the frames every ~20 ns. And this is not what we want.
The fix consists in taking a lock before transmitting the frame, both
for manual and fdma injection. In this way the DCB chain would not be
closed. The issue is not seen on lan969x or lan966x because there we
have only 1 core.