net: lan966x-pci: add API for configuring PCI ATU
Add a small API for configuring the PCI ATU (Address Translation Unit).
The ATU makes it possible to divide the PCI outbound window into
translateable regions. This is used for FDMA-based RX/TX, as the
addresses used by the FDMA will have to be mapped inside the OB window.

This naive implementation divides the OB window into 6 equally sized
regions.

Writing to the ATU registers is done directly using writel() instead of
lan_wr() and friends. This is because the 6 ATU region register is not
indexed, but rather prepended an index number - eg. region 5 is called
IATU_LWR_BASE_ADDR_OFF_OUTBOUND_5. So to avoid a bunchy of boilerplate
code, the offsets are defined internally and used to access the correct
registers.