Revert "sparx5: Take tx lock when injecting frames"
This reverts commit b3470914e65dd457bf732d3c1bcad13b2070d791.

It was noticed that the boards were rebooted randomly when the APPL was
running idle on the lan969x board. And it seems that this issue was
introduced recently. So to help the APPL try to revert this commit as
this takes a lock and we think there is a deadlock.