sparx5: Fix use after free
This issue was reported by a customer who enabled CONFIG_KFENCE and then
created a bridge and add some ports under the bridge and make sure that
in the MAC table were added some dynamic entries. Once these entries
expired then the following warning has been seen:

[  213.406814] ==================================================================
[  213.413732] BUG: KFENCE: use-after-free read in sparx5_del_mact_entry+0xf4/0x110
[  213.413732] [  213.422238] Use-after-free read at 0x(____ptrval____) (in kfence-#112):
[  213.428629]
[  213.430219]
[  213.431646] CPU: 1 PID: 8 Comm: kworker/u4:0 Not tainted 6.1.55 #2
[  213.437559] Hardware name: Sparx5 PCB134 Reference Board (NAND) (DT)
[  213.443636] Workqueue: sparx5_order sparx5_switchdev_bridge_fdb_event_work
[  213.450221] ==================================================================

I simple couldn't reproduce the issue but based on static analyzes of
the code it looks like mact_entry is used after being freed. The fix
consists in changing the order of operations such that mac entry is
removed and after that free the memory