net: sparx5-serdes: Fix access to SD_LANE_25G
When booting sparx5 it was noticed a crash in the serdes. And the reason
is because it was accessing some registers that were not mapped. And
they were not mapped because they don't exist.
So the issue was that it was calculating wrong the number of SD_LANE_25G.
There are only 8 instead of 12. With this change, it stops crashing the
kernel.