net: sparx5-appl: Fix checking of sparx5
With the current code, all the versions of sparx5 and lan969x are using
the FDMA to inject/extract frames. This is not correct because first
revision of sparx5 doesn't support fdma.
The issue that the function is_sparx5 will return always fals when
called in the sparx5-appl therefore will aways use the FDMA. And will
return false because sparx5->target_ct is not updated, it has a default
value of 0.
The fix consists in updating the sparx5->target_ct with the correct part
id.