sparx5: fix reconfiguration of PCS on link mode change
It was observerd that SFP's was broken on Sparx5 PCB134 and have been
since commit: 6e5bb3da9842950a161625b4ab2743d1d5c64715.

The issue is observed when using copper SFP's and the link mode changes
from not having inband support, to having inband support - and vice
versa. The reason, is because the inband flag of the old config is used,
instead of the inband flag of the new config.

This used to work, as the inband flag of the old config was '1', but
with the new changes of said commit, the flag is now '0', as it ought to
be (at least that is my understanding, that 10GBase-R does not support
inband negotiation), causing wrong configuration.

Optical SFPs are not impacted by the bug.