net: sparx5: lan969x: Fix the page_order for APPL
When the APPL is used, it would set the mtu of the device to be 10400
but it would allocate page of 4096 for the received frames. That means
when receiving a frame bigger than 4096 it would fail to add it to the
skb as the skb is build around the page, so it would crash.
Fix this by allocating a page of order 2, as it needs to be bigger than
10400. Therefor a order of 2 is chosen.
This affects only the APPL for lan969x, as sparx5 is not used this
page_order when allocating the received frames.