net: sparx5: sparx5: Update the tx side to work APPL
When the APPL is used, will set the mtu to a value of 10400. But the
buffer allocated for the transmit frames was set to 2048. Meaning that
whenever we tried to send a frame bigger than that, it would failed to
transmit correctly the frame. The fix consists in using the device mtu
to see how big the buffer needs to be.