net: sparx5: sparx5: Remove hardcoded value for xtr buffer
The issue is that if a frame bigger than 2048 arrives to the CPU then
there will be a panic. This is possible only when the fdma is used with
the APPL because the APPL creates only 1 network device and sets the mtu
to be 10400. To fix this, allocate frames big enough to received the
biggest frame (10400) and for this use the rx->page_order. As this is
also used by the lan969x we can use this and this is already set to a
value of 2.
This change also affects the standalone kernel but in that case the
rx->page_order has a value of 0, so now we frames that has a size of
4096 instead of 2048. But this is one of the first steps to actually be
able to change dynamically the mtu of the device for standalone because
currently max value is 1500.