net: sparx5: vcap: fix uninitialized addr detection
Key and mask entries read from uninitialized VCAP super addresses
consist of contiguous 1's for the full length of the address in HW, e.g.
52bits.

When read into u32 words, the additional bits in the final word contain
0. For instance with 52bit addresses we need 2 u32 words so the final
word looks like: 00000000000011111111111111111111.

Therefore, we need to mask the final word, for the == 0 check for work.
Note that sw_width is not necessarily equal to the full HW width, since
there may be waste. On lan969x the HW width is 52, but LPM sw_width is
just 49.

However, for the purposes of detecting uninitialized addresses it should
be sufficient to use a mask based on sw_width, as nothing can be written
to the additional bits we mask out.

The reason this surfaces now, is that the uninitialized values was
filtered at a later stage in vcap_verify_keystream_keyset. However, the
first lpm keyset has type_id == -1, in which case
vcap_verify_keystream_keyset always returns true.