uio: sparx5_irqmux: Fix resource iterations
The driver will create a map for each of the resources that it has. The
problem is that it iterates wrong the resources, it goes beyond the end
of the resource vector in case the bits after the resources matches to
be a memory resource type. In that case it would add another map. Until
it doesn't find a memory resource type. Another issue with this could be
that the first resource type not to be memory but something else, then
it wouldn't add any maps. It was noticed that 1 of 1000 times actually
the driver was creating 3 maps instead of 1 as set in device tree.

The fix consists in actually checking how many resources are and then
iterate all of them and then create maps only for memory resources.