dts: mips: ocelot: Remove the gpio interrupt from uio device
On ocelot compared to the other mips platforms, the GPIO interrupt is
used by the gpio device. As there are phys that generated an interrupt
that goes to the gpio controller.
Therefore this interrupt can't go at the same time to two devices (one
the gpio and one to uio device) so remove the interrupt from the uio
device.