sparx5: Fix when setting management port down
The following commands resulted in a crash:
ip link set dev eth29 up
ip link set dev eth29 down

The reason is that when the port is set down, it would check if there is
any PHY attached and if not, it would set the serdes down. But before
checking if there is a PHY the PHY is disconnected meaning that will
always not be a PHY at the point of the check.  So it would always try to
configure the serdes. But there might be ports (like management) that
don't have a serdes so then there is no serdes to configure.
Therefore change the check, to check for serdes instead of phy which
will not be there.