net: microchip: fdma: add minimal platform agnostic FDMA API
The FDMA hardware block is the same across lan966x, lan969x and sparx5
platforms. Lets take advantage of this, and gather up all the different
approaches into one FDMA library.

The API revolves around a single fdma struct, that holds all the
cookkeeping data for a single RX or TX channel. The user will preset the
size of the contiguous DMA space and the size of each data buffer pointed
to by the dataptr variable.

Adding DCB's to the chain is done by the fdma_dcb_add() function, which
will use the preset nextptr() or dataptr() pointers, or by using the
fdma_dcb_add_cb() that takes the pointers as arguments. I found it
necessary to add this indirection, in order for the users *not* to have
to loop through DCB's or DB's in order to set these pointers manually.
Instead the user will implement the nextptr() and dataptr() functions,
and the library will call it to retrieve the correct addresses.

Some clarification on naming:

fdma_alloc() and fdma_free() will only allocate the dma space from the
dma mapping api function dma_alloc_coherent()

fdma_alloc_and_map() and fdma_free_and_unmap() will do likewise, but at
the same time map the dma space to the PCI ATU.

fdma_alloc_phys() and fdma_free_phys() will allocate the dma space from
kzalloc() and convert the virtual address to a physical address using
virt_to_phys().

Theres a number of functions ending in *_contiguous(). These are for use,
when the dataptr buffers reside in the same memory segment as the DCB's,
as is the case with the PCI drivers and sparx5.

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>