net: sparx5: vcap: fix lpm sw_width
The XML file used to autogenerate VCAP API files, has an error.
It lists sw_width as 49, instead of 52. sw_width is supposed to be
the maximum used bits for an addresss, including type group and type id
bits.
A single LPM address is 52 bit wide (SUPER VCAP), and there are LPM keys
which used more than 1 address. Therefore, the sw_width must be 52,
similar to the other VCAP's using the super blocks.

The sw_width is also used to calculate the field offsets for keys, so
these must be corrected as well.