net: sparx5: create helper function for configuring LLP registers
Create a helper for configuring the LLP registers. This will be needed
later when introducing support for FDMA over PCI. Also skip the
configuration of the LLP registers in rx/tx_activate, as that assumes
the base address is the dma address, which is not always the case (e.g
PCI solution where the DMA addresses are mapped).

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>