net: sparx5/lan969x: let xmit functions consume SKB
Currently, it is assumed that the SKB should be consumed if not a PTP
frame. This will no longer work, when the lan969x driver will transition
to TX cleanup in the RX loop. So instead, let the specific xmit
function decide if it wants to consume the SKB or not.

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>