net: lan969x: use FDMA API on the TX side
Make use of the FDMA API in all TX related functions (primarily alloc,
free, xmit ..) *and* use the DMA mapping API, as the driver ought to do.

The driver did not use the FDMA API as it ought to do, because of,
reasons. Now use the DMA mapping API to map the SKB to DMA space and
pass the dma addr as the dataptr of the DB's. This also requires that
the SKB's are unmapped and consumed on the RX side, thus requiering a
bit of bookkeeping code in sparx5_db. The actual unmapping and consuming
will be implemented on the RX side in a following commit.

Also get rid of first_dcb and curr_dcb as these pointers were
placeholders and are no longer required.

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>