ptp: zl3073x: Fix freq adj when running on 32bit arch.
The function zl3073x_ptp_adjfine contains the value with how much the
clock frequency needs to be adjusted. The type of the argument is a long
and on arm32 this has a size of 4 bytes. The problem is that it can
happen such that the multiplication between the scaled value and the
1PPM format to be bigger than 32bit which means that the value will wrap
around so it would give a wrong value.
The fix consists in making sure to calculate this using s64. This seems
to be fixing the issue.