net: sparx5: Fix l3 forwarding ipv6 blackhole offload detection
In the kernel IPv4 and IPv6 differs w.r.t to the egress device used
for nexthops of blackhole, prohibit and unreachable route types.

For IPv4 the egress device is NULL, but for IPv6 the loopback interface
is used.

The IPv6 case would fail the egress-device-is-vlan-aware check. We fix
this by returning early on these route types.