net: sparx5: Fix mactable sleep in atomics with spinlock
The sleep-in-atomics flag caught a problem with sparx5_set_rx_mode is
called in atomic context, and sparx5_mc_*sync takes the mac register
mutex and does a wait for completion.

This changes mact_wait_for_completion to use the atomic version, and
replaces the mac register mutex with a spinlock.