net: lan966x: Timestamp all frames when timestamping is enabled
Currently when HW timestamp is enabled for transmitted frames then it
would just timestamp only PTP frames. And that works fine. But
apparently there are cases where all the transmitted frames need to be
timestamped (2-step) in TSN network. Therefore change this to timestamp
all the frames when it is enabled and not only the ptp frames.