dts: lan969x_appl_*.dts: Fix interrupt for oam_vop
Currently the oam_vop interrupt was mapped to the analyzer interrupt.
But it is supposed to be mapped to oam_mep interrupt. We had this
mistake because this was copy from lan966x which in that case the
oam_mep interrupt is under the analyzer. So the fix consists in using
the correct interrupt, this is similar to sparx5.