net: microchip: vcap: Fix out-of-bounds reads
Three tests exhibited out-of-bound reads, flagged by KASAN:

1) vcap_api_addr_keyset_test
2) vcap_api_show_admin_raw_test
3) vcap_api_iterator_init_test

For 3), the issue is the test variable typegroups2 does not end with a
zero struct. Yet, vcap_iter_skip_tg assumes typegroup arrays end with a
zero struct for the loop condition. This is a snippet of the KASAN
report:

BUG: KASAN: stack-out-of-bounds in vcap_iter_init+0x86/0x120
Read of size 2 at addr 00000000629f3d24 by task kunit_try_catch/30
...
The buggy address belongs to stack of task kunit_try_catch/30
 and is located at offset 116 in frame:
 vcap_api_iterator_init_test+0x0/0x540
...

Since the offset of 134 is larger than the last tg in typegroups2, at
offset 98, this causes vcap_iter_skip_tg to iterate beyond typegroups2,
in the call:

	vcap_iter_init(&iter, 49, typegroups2, 134);

For 1) and 2) the issue is that reading a VCAP addresss with
vcap_addr_keysets, will read multiple words from the key- and maskstream
buffers. Here are snippets of the two KASAN reports:

BUG: KASAN: stack-out-of-bounds in test_cache_read+0xcb/0xf0
Read of size 4 at addr 00000000625fbc60 by task kunit_try_catch/18
...
The buggy address belongs to stack of task kunit_try_catch/18
 and is located at offset 336 in frame:
 vcap_api_addr_keyset_test+0x0/0x4d0
...

BUG: KASAN: stack-out-of-bounds in vcap_verify_typegroups+0x188/0x250
Read of size 4 at addr 00000000625fbb18 by task kunit_try_catch/20
...
The buggy address belongs to stack of task kunit_try_catch/20
 and is located at offset 56 in frame:
 vcap_api_show_admin_raw_test+0x0/0x540
...

First of all, the bit-width of address is larger than 32. The specific
size depends on the VCAP type, but for IS2 the width is 52 bits.
Therefore, reading a single VCAP address will read 2 u32 words from the
buffers in read_cache. This is the reason we can not start iteration at
ARRAY_SIZE(keydata) - 1, but must subtract rd_ahead.

Moreover, the subsequent decoding of the keyset typegroups will read as
many words as necessary to validate the typegroups. IS2 has an X12 rule
with the final typegroup offset at bit 468, which means we may read 18
positions from the key and mask stream buffers.

However, we have only read a single address into the key/mask buffers,
so we are decoding data in the buffers which has not been read from hw.
These are just all zero. The reason this worked at all, and was able to
detect keysets is that typegroup values, after the first one, are always
a sequence of zero bits.

Changing sw_max from 0 to 1 ensures we only look at the first typegroup,
when we have only loaded one address from hw.

We can fix this either by increasing the buffer size in the test, and
adjusting the starting read position, or by changing sw_max from 0 to 1
in the vcap_find_keystream_keysets call, ensuring we only validate the
first typegroup.

Flagged by KASAN.