net: microchip: vcap: remove incorrect err check in debugfs
The error rule printing will cause buggy debugfs output in some cases.
Since we are looking at a single VCAP address at a time, and trying to
infer if it contains the start of a rule.

It is possible for addresses in the middle of a rule, to match the
typegroup bits of some other keyset/entry. For instance this entry:

vcap add 1 lpm 10 0 vcap_kfs_dbl_ip6 affix 0x8f 0xfff \
     ip6_dip 0xf0f0f0f0f0f0f0f0ffffffffffffffff \
             0xffffffffffffffffffffffffffffffff \
     ip6_sip 0xf0f0f0f0f0f0f0f0ffffffffffffffff \
             0xffffffffffffffffffffffffffffffff \
     vcap_afs_l3mc_ptr l3mc_ptr 15

I am not sure why I did not catch this earlier, as I used the debugfs
extensively. However, since we now show all raw addresses we should just ignore the
case where addr % info->sw_per_item > 0, and show it as a raw address.

Worth noting this method can give false positives. Properly detection of
a keyset requires reading multiple addresses, and then inspecting all
typegroup bits and type_ids.

We are reading a single address at a time, and only check that the first
typegroup bits match, and that the current address aligns to the size of
the identified keyset.